Moved Vitalybaev\GoogleMerchant\Product\Meta() to Vitalybaev\GoogleMerchant\Meta\Product()
